Adds Get-B3StringHash function to module.
Version 0.2.1:
New function has a help entry and the ability to create a hash of a string value and returns just the hash as a string (same as b3sum.exe) instead of a custom PowerShell object like hashing a file. Key derivation is still not implemented.
All tests pass, except for the one realted to key derivation.